Using integers to return contours points is better than using double.
git-svn-id: file:///home/adenilson/temp/wtf/repository@170 56e0a439-041b-0410-8790-949b9ad245cc

+++ thesis/utest/aux_test.cpp
@@ -119,7 +119,7 @@ START_TEST (t_adapt_curvature)
 	CvSeq *sequence = NULL;
 	int num_contours, i;
 	CvMemStorage* storage = cvCreateMemStorage(0);
-	ocv_adaptor<double> handler;
+	ocv_adaptor<int> handler;
 	double *curvature = NULL;
 	double tau = 10.0, c_energy;
 
@@ -127,7 +127,8 @@ START_TEST (t_adapt_curvature)
 	handler.reset(sequence);
 
 	do {
-		curvature = contour_curvature<ocv_adaptor<double>, mcomplex<double> >(handler, handler.contour_length(), tau);
+		curvature = contour_curvature<ocv_adaptor<int>,
+		  mcomplex<double> >(handler, handler.contour_length(), tau);
 		c_energy = energy(curvature, handler.contour_length());
 		delete [] curvature;
